lotrs-py/requirements.txt
numpy>=1.24
pycryptodome>=3.19
mpmath>=1.3
